Implement basic chmod() syscall and /bin/chmod helper.
Only raw octal modes are supported right now.
This patch also changes mode_t from 32-bit to 16-bit to match the on-disk
type used by Ext2FS.

I also ran into EPERM being errno=0 which was confusing, so I inserted an
ESUCCESS in its place.

+++ LibC/sys/types.h
@@ -24,7 +24,7 @@ typedef uint32_t ino_t;
 typedef int32_t off_t;
 
 typedef uint32_t dev_t;
-typedef uint32_t mode_t;
+typedef uint16_t mode_t;
 typedef uint32_t nlink_t;
 typedef uint32_t blksize_t;
 typedef uint32_t blkcnt_t;
